requirements.txt
requests>=2.31.0
sympy>=1.12
python-dotenv>=1.0.0
